Rename pluginContext to constants
